+++ sql/mysql/tables-generated.sql
@@ -10,3 +10,13 @@ CREATE TABLE /*_*/pagelike_like (
   INDEX pagelike_like_user_page (pll_user_id, pll_page_id),
   PRIMARY KEY(pll_page_id, pll_user_id)
 ) /*$wgDBTableOptions*/;
+
+
+CREATE TABLE /*_*/pagelike_notification_dedupe (
+  plnd_page_id INT UNSIGNED NOT NULL,
+  plnd_user_id INT UNSIGNED NOT NULL,
+  INDEX pagelike_notification_dedupe_user_page (
+    plnd_user_id, plnd_page_id
+  ),
+  PRIMARY KEY(plnd_page_id, plnd_user_id)
+) /*$wgDBTableOptions*/;